Use a QScrollArea for the settings widget
This commit add a scroll area for the settings widget to it can also be
used with small window sizes. Only a vertical scroll bar is allowed
(horizontal spacing shouldn't be a problem). Replace the fixed size
definition with a sizing policy to enable the settings area to grow in
vertical direction if possible.